feat: split newtab specs and enable parallel E2E execution
- Split newtab.spec.ts into 5 focused test files:
  - newtab-display.spec.ts: 基本表示 (NEW-001, NEW-002, NEW-003, NEW-008)
  - newtab-stats.spec.ts: 統計カード (NEW-004, NEW-013)
  - newtab-goal-edit.spec.ts: 目標編集 (NEW-006, NEW-007)
  - newtab-block-info.spec.ts: ブロック情報表示 (NEW-009, NEW-010, NEW-012)
  - newtab-settings.spec.ts: 設定とプレミアム機能 (NEW-005, NEW-011)
- Enable parallel execution in playwright.config.ts:
  - workers: 3
  - fullyParallel: true
- Remove unused imports from PR #274 cleanup:
  - analytics.spec.ts: removed setStorageData, clearStorage, getStorageData
  - block.spec.ts: removed setStorageData, clearStorage
  - data.spec.ts: removed setStorageData, getStorageData, clearStorage
  - i18n.spec.ts: removed setStorageData, clearStorage
  - interaction.spec.ts: removed setStorageData, clearStorage
